// apps/nfid-frontend/src/integration/_ic_api/internet_identity.d.ts
export type CredentialId = Array<number>
export interface Delegation {
  pubkey: PublicKey
  targets: [] | [Array<Principal>]
  expiration: Timestamp
}
export interface DeviceData {
  alias: string
  protection: DeviceProtection
  pubkey: DeviceKey
  key_type: KeyType
  purpose: Purpose
  credential_id: [] | [CredentialId]
}
export type DeviceKey = PublicKey
export type DeviceProtection = { unprotected: null } | { protected: null }
// ...
export type KeyType =
  | { platform: null }
  | { seed_phrase: null }
  | { cross_platform: null }
  | { unknown: null }
export type PublicKey = Array<number>
export type Purpose = { authentication: null } | { recovery: null }
// ...
export type UserNumber = bigint